test: split up the maldisplay tests
